Add object Equality tests
