+++ src/ValidatableExtensions/ValidatableExtensions.Comparables.cs
@@ -38,38 +38,6 @@ public static partial class ValidatableExtensions
     }
 
     /// <summary>
-    /// Throws an exception if the value equals <paramref name="n"/>.
-    /// </summary>
-    /// <remarks>
-    /// The default exception thrown is an <see cref="ArgumentException"/>.
-    /// </remarks>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static ref readonly Validatable<TValue> IfEquals<TValue>(this in Validatable<TValue> validatable, TValue n)
-        where TValue : notnull
-    {
-        Validator.ThrowIfEquals(validatable.Value, n, validatable.ParamName, validatable.ExceptionCustomizations);
-
-        return ref validatable;
-    }
-
-    /// <summary>
-    /// Throws an exception if the value does not equal <paramref name="n"/>.
-    /// </summary>
-    /// <remarks>
-    /// The default exception thrown is an <see cref="ArgumentException"/>.
-    /// </remarks>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static ref readonly Validatable<TValue> IfNotEquals<TValue>(
-        this in Validatable<TValue> validatable,
-        TValue n)
-        where TValue : notnull
-    {
-        Validator.ThrowIfNotEquals(validatable.Value, n, validatable.ParamName, validatable.ExceptionCustomizations);
-
-        return ref validatable;
-    }
-
-    /// <summary>
     /// Throws an exception if the value is greater than 0.
     /// </summary>
     /// <remarks>
